API_applicationProject/requirements.txt
# pip install -r requirements.txt to get packages
argparse
google-generativeai
google-cloud-aiplatform
firebase-admin
bcrypt
python-dotenv